option 2 completed for addon
