+++ tests/examples/users.js
@@ -1,0 +1,16 @@
+const schema = {
+  properties: {
+    name: { type: 'string' },
+    email: { type: 'string' },
+    password: { type: 'string' }
+  },
+  required: ['name', 'email', 'password']
+}
+const example = {
+  name: 'page',
+  email: 'page@gmail.com',
+  password: 'pagethesdwrapper'
+}
+module.exports = {
+  schema, example
+}
