+++ tests/quaterCup.test.js
@@ -10,25 +10,8 @@ describe('this test prevents to any issues and problems, also to break the struc
     expect(staticData.quaterCup).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    name: { type: 'string' },
-    americanStandart: { type: 'string' },
-    americanStandartOunces: { type: 'string' },
-    volume: { type: 'string' },
-    weight: { type: 'string' }
-  },
-  required: [ 'name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight' ]
-}
-
-const example = {
-  name: 'teaspoon',
-  americanStandart: '4 tablespoons',
-  americanStandartOunces: '2 ounces',
-  volume: '2 milliliters',
-  weight: '56.7 grams'
-}
+const schema = require('./examples/quaterCup').schema
+const example = require('./examples/quaterCup').example
 describe('test quarter cup json schema', () => {
   it('validates my json', () => {
     expect(example).toMatchSchema(schema)
