+++ tests/teaspoon.test.js
@@ -10,25 +10,8 @@ describe('this test prevents to any issues and problems, also to break the struc
     expect(staticData.teaspoon1).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    name: { type: 'string' },
-    americanStandart: { type: 'null' },
-    americanStandartOunces: { type: 'null' },
-    volume: { type: 'string' },
-    weight: { type: 'null' }
-  },
-  required: [ 'name', 'americanStandart', 'americanStandartOunces', 'volume', 'weight' ]
-}
-
-const example = {
-  name: 'teaspoon',
-  americanStandart: null,
-  americanStandartOunces: null,
-  volume: '2 milliliters',
-  weight: null
-}
+const schema = require('./examples/teaspoon').schema
+const example = require('./examples/teaspoon').example
 describe('testng for teaspoon json data schema', () => {
   it('validates my json', () => {
     expect(example).toMatchSchema(schema)
