+++ tests/types.test.js
@@ -12,21 +12,8 @@ describe('types data files returns array', () => {
     expect(types).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    Field: { type: 'string' },
-    Type: { type: 'string' },
-    Description: { type: 'string' }
-  },
-  required: ['Field', 'Type', 'Description']
-}
-
-const example = {
-  Field: 'from',
-  Type: 'integer',
-  Description: 'First result index as submitted'
-}
+const schema = require('./examples/types').schema
+const example = require('./examples/types').example
 
 describe('types json schema testing', () => {
   it('validates type json schema', () => {
