+++ tests/userGrocery.test.js
@@ -11,18 +11,8 @@ describe('userGrocery data files returns array', () => {
     expect(userGrocery).not.toBe('')
   })
 })
-
-const schema = {
-  properties: {
-    user_id: { type: 'string' },
-    grocery_id: { type: 'string' }
-  },
-  required: ['user_id', 'grocery_id']
-}
-const example = {
-  user_id: 'fbe57252-7a3e-11e9-8ec5-0d5ec94f7bcf',
-  grocery_id: 'fbe57236-7a3e-11e9-8ec5-0d5ec94f7bcf'
-}
+const schema = require('./examples/userGrocery').schema
+const example = require('./examples/userGrocery').example
 describe('testing userGrocery schema and also requiring it', () => {
   it('validates UserGrocery json schema', () => {
     expect(example).toMatchSchema(schema)
